feat: remove GameFactory and migrate to static factory methods in Game (fixes #7)
