+++ src/App.tsx
@@ -2,12 +2,12 @@ import React, { useMemo } from 'react';
 import { CanvasView } from './canvas/components/CanvasView';
 import { Session } from './models/Session';
 import { User } from './models/User';
-import { GameFactory } from './models/GameFactory';
+import { Game } from './models/Game';
 
 function App() {
   const session = useMemo(() => {
     const user = new User('local-player');
-    const game = GameFactory.createStandardGame();
+    const game = Game.createStandard();
     const s = new Session('local-session', user);
     s.startNewGame(game);
     return s;
